Increase maxFrameBufferAcquiredBuffers to 3
When display and SF falls out of sync, there is a possibility that while
display holds on to a buffer, a composed buffer could be waiting in the
decon thread. Since maxFrameBufferAcquiredBuffers is 2, GPU will have to
wait for display to swap out the buffer it is currently presenting to be
able to compose for the next frame, which leads to SF missing its
timeline for subsequent frames.

Bug: 190842189
Test: perfetto

Signed-off-by: Devika Krishnadas <kdevika@google.com>
Change-Id: I434e8644cff2538589c1c9b2bf731afae0aa8980