gs101: Use AOSP build framework for pvmfw.img
Remove the definition for BOARD_PREBUILT_PVMFWIMAGE (which currently
points to the shared AOSP pre-built of pvmfw.img) to fall back to the
default behavior of the AOSP framework. Even though this will result in
the same pre-built being used (for now), not setting that variable still
partially modifies the overall behavior (e.g. for packaging) and allows
upcoming framework changes to be automatically picked up.

Bug: 199717422
Test: m
Change-Id: I7c3bbbee8c83d262b707f8582b5f43040bc25a06