Remove ndk_platform backend. Use the ndk backend.
The ndk_platform backend will soon be deprecated because the ndk backend
can serve the same purpose. This is to eliminate the confusion about
having two variants (ndk and ndk_platform) for the same ndk backend.

The prebuilt files had a reference to
android.hardware.gnss-V1-ndk_platform.so. Change it to
android.hardware.gnss-V1-ndk.so by diretly editing the DT_NEEDED entry.

Bug: 161456198
Test: m
Change-Id: I25ec19101c408cc29df38d127f284c7085fe1d1a