wifi: Enable ieee80211w feature on SoftAp interface
Set ieee80211w=1 config item in the hostapd config file
based on the feature flag, BOARD_HOSTAPD_CONFIG_80211W_MFP_OPTIONAL.
This enables management frame protection and prevents denial of
service when an attacker sends deauthentication frames to the
clients connected to the SoftAp.

Bug: 197874458
Test: Connected STAs to Pixel SofAp and confirmed from the
sniffer logs that PMF capabilities are exchanged and the required
management frames are protected.

Change-Id: I6f23fd82eeec97d80d808c90ad48901eda630646