FUSE-BPF: force disable sysprop flag
The FUSE-BPF feature has been enabled on some testing devices through
the system property flag "persist.sys.fuse.bpf.enable", but the flag
enabling change has been reverted as soon as failures has been reported.
Unfortunately, being the system property persistent, reverting the
enabling change is not sufficient to turn the feature down on devices
which already have it enabled, for which it has to be manually set to
"false".

Bug: 221812566
Test: getprop persist.sys.fuse.bpf.enable
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I87c93ddaefd79dfcf900eb2cbe2fe7518bf24755