Use the ?= operator so BT properties can be overridden
System properties set with the "=" operator cannot be overriden. This
device is used elsewhere in the system and some of the properties need
to be overridden. The "?=" operator will use the assigned value if no
other values are assigned elsewhere, allowing for overrides.

Bug: 230638462
Test: adb shell getprop | grep bluetooth
Merged-In: I45b1786f5556ba3e7d493ea17dc955cd317a1156
Change-Id: I45b1786f5556ba3e7d493ea17dc955cd317a1156