Import prebuilt kernel modules like fips140.ko
Add support for copying prebuilt kernel modules like fips140.ko into the
vendor ramdisk. These kernel modules are different in the following
ways:

- They do not appear in vendor_boot.modules.load provided by the
  kernel build system.
- We want to put their names at the very beginning of the output
  modules.load file to ensure that they are loaded prior to all other
  modules.

We define a new Makefile variable named
BOARD_PREBUILT_VENDOR_RAMDISK_KERNEL_MODULES. For every filename on this
list, we verify that the file exists in $(KERNEL_MODULE_DIR), and if it
does, we include its name in the modules.load for the vendor boot
ramdisk.

Bug: 194712883
Change-Id: Iff8923c2492296b8eb4d24ec94bbfccc19f4e8f5