dumpstate: add HWC fence state, error log, and debug dump files
Add some HWC log files in /data/vendor/log/hwc to dumpstate, i.e.

[1] hwc_fence_state.txt: contains the dump of fence state when there's
a fence leak warning. We can also execute `vndservice call
Exynos.HWCService 11 i32 0 i32 308 i32 1` to trigger the dump to this
file. The size of this file is limited to about 1 MB.

[2] hwc_error_log.txt: contains some early logs of ALOGE() from HWC.
The size of this file is limited to about 1 MB.

[3] {display_name}_hwc_debug0.dump - {display_name}_hwc_debug9.dump:
contain the debug dumps when there are some failures in HWC (e.g.
failures in validateDisplay() or presentDisplay()). HWC will only limit
up to first 10 debug dumps per display to be saved in separated files.

Bug: 230361290
Test: adb bugreport
Change-Id: If2158a326faef96c0e0cd4e93a17ec420409bf1e