Update the i2c device node sepolicy labeling
This change needs to be merged with the corresponding kernel change that
sets the i2c bus aliases correctly to match the existing v5.10 bus
probe ordering.

To verify the sepolicy labeling doesn't change, run the below commands
and diff the stdout on builds with and without the changes. For extra
credit, verify the nodes are labeled the same when upgrading the kernel
to v6.1 (with the correct i2c aliases to match the existing policy):

  acpm_bus_array=("acpm_mfd_bus@17500000" "acpm_mfd_bus@17510000")
  for bus in ${acpm_bus_array[@]}; do
    adb shell ls -ZR /sys/devices/platform/${bus}/i2c-*;
  done

  bus_array=("10960000" "10970000" "10d50000" "10900000")
  for bus in ${bus_array[@]}; do
    adb shell ls -ZR /sys/devices/platform/${bus}.hsi2c/i2c-*;
  done

Test: verify on r4
Bug: 291606723
Change-Id: Ifbfc53fbeb39a47cda4263fc706f11af6675d90e