liboemservice_proxy: Remove service defs
`Liboemservice_proxy` was originally included as a part of the build
target for DMD, which meant that the binder aidl declaration had to be
done on all devices that used DMD (all Lassen devices), even if they
didn't use the service. `Liboemservice_proxy` has now been decoupled
from DMD so it only needs to be included on our target devices now. This
CL removes the outdated service definitions.

Test: Build gs101 and ensure it still builds.
Bug: 321867236
Change-Id: Iefd8a4e563c41ffc5221810e3fb042dbe4a3e733