Disable dumpstate strict-run flag in gs101
The change disables the dumpstate strict-run flag in gs101 devices, which is enabled by default. Disabling this flag increases the timeout for capturing incident report to 120 seconds, from the default 20 seconds. This is done because 20s isn't enough to capture incident report proto in some cases, especially when device is throttling.

Bug: 330400782

Test: Booted a gs101 device with the change. Then manually thermal throttled the device and captured bugreports. In some tests Incident report took a lot more more than 20s, and in all these cases it was succesfully captured.
Change-Id: I8c4ee4e4412c517822776a9b10bbdad7da1c1213